chore(frontend): move shadcn imports
