+++ apps/frontend/components.json
@@ -14,7 +14,7 @@
     "aliases": {
         "components": "@/components",
         "utils": "@/lib/utils",
-        "ui": "@/components/ui",
+        "ui": "@/components/primitives",
         "lib": "@/lib",
         "hooks": "@/hooks"
     },
