+++ apps/frontend/src/components/primitives/input-group.tsx
@@ -2,9 +2,9 @@
 
 import { cva, type VariantProps } from "class-variance-authority";
 import type * as React from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
+import { Button } from "@/components/primitives/button";
+import { Input } from "@/components/primitives/input";
+import { Textarea } from "@/components/primitives/textarea";
 import { cn } from "@/lib/utils";
 
 function InputGroup({ className, ...props }: React.ComponentProps<"div">) {
